pyproject.toml
[project]
name = "notion-mcp"
version = "0.1.0"
description = "Notion MCP Server - Integrate Notion calendar and list databases with Model Context Protocol"
authors = [
    { name = "User", email = "user@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "notion-client>=2.2.1",
    "python-dotenv>=1.0.0",
]

[project.scripts]
notion-mcp = "notion_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []
